requirements.txt
PyQt5~=5.15.10
python-dotenv~=1.0.0